+++ src/lib/utils/candidatePipeline/candidatePipelineFormatters.js
@@ -1,7 +1,7 @@
 export function getCurrentTimestamp() {
   return new Date().toLocaleString("en-PH", {
     year: "numeric",
-    month: "short",
+    month: "long",
     day: "numeric",
     hour: "numeric",
     minute: "2-digit",
@@ -16,7 +16,7 @@ export function formatDateTime(date) {
   if (!date) return "—";
 
   return new Date(date).toLocaleDateString("en-PH", {
-    month: "short",
+    month: "long",
     day: "numeric",
     year: "numeric",
     hour: "numeric",
